Move: 関数を分割して役割を明確化
